+++ src/components/index.ts
@@ -1,2 +1,8 @@
 export {default as Header} from "./common/Header";
 export {default as Footer} from "./common/Footer";
+export {default as Brokers} from "./home/Brokers";
+export {default as TopBrokers} from "./home/TopBrokers";
+export {default as BrokerLogo} from "./home/BrokerLogo";
+export {default as BrokerName} from "./home/BrokerName";
+export {default as Search} from "./home/Search";
+export {default as RequestSource} from "./RequestSource";