+++ src/types/common.type.ts
@@ -1,0 +1,6 @@
+
+export interface IResponse<T> {
+    data: T;
+}
+
+export type REQUEST_SOURCE_TYPE = "api" | "local";