+++ change-log.md
@@ -1,12 +1,31 @@
-## <u>Broker Chooser</u> 
+## <u>Broker Chooser</u>
 
 ---
 
 | Version            |
 |:-------------------|
+| [`0.3.0`](#v0.3.0) |
 | [`0.2.0`](#v0.2.0) |
 | [`0.1.0`](#v0.1.0) |
 
+
+### <a id="v0.3.0"/>`0.3.0`
+
+- Added `config` folder and `API.ts` to configure my requests
+- Added `services` to handle my services and dependency injection
+- Added `utils` folder to use helper functions
+- Change `next.config.js` to support the image URL
+- Added `exteraReducers` to make a `Async` request
+- Added `RequestSource.tsx` component to change the source of the application
+  - `Local` to use directly from the `data/brokers.ts`
+  - `Api` to use the APIs that provided in `pages/api`
+- Added `useSearch` hook to handle and organized the search logic
+- Added `useBrokers` and `useTopBrokers` hooks to handle their logic's in one place
+- Added `debounce` helper function to optimized function call
+- Added `sendEvent` logic to the app (the basic logic)
+- Added responsive styles to the app
+- Added `react-intersection-observer` to check the element in the viewport
+- Added `useImpression` hook to check the observed element
 
 ### <a id="v0.2.0"/>`0.2.0`
 
@@ -18,6 +37,7 @@
 - Added `Header` and `Footer` to make the app scalable
 - Changed `_app.tsx` and `index.tsx` in the `pages` folder
 - Added `react-icons` to use in the app
+- Added `Volta` for checking `Node` version to the `package.json`
 
 ### <a id="v0.1.0"/>`0.1.0`